refactor: rename `area-boundary.tsx` into `AreaBoundary.tsx`
