+++ components/AreaBoundary.tsx
